Make simu pgm addr 64 bits is better ... on 64bits archs

+++ branches/frsky/src/simpgmspace.h
@@ -30,7 +30,7 @@ typedef const int8_t prog_int8_t;
 #define PROGMEM
 #define pgm_read_byte(address_short) (*(uint8_t*)(address_short))
 #define pgm_read_word(address_short) (*(uint16_t*)(address_short))
-#define pgm_read_adr(address_short) (*(int*)(address_short))
+#define pgm_read_adr(address_short) (*(int64_t*)(address_short))
 #define pgm_read_stringP(adr) ((adr))
 #define PSTR(adr) adr
 #define _delay_us(a)
